Split ghclient into core and wrappers.
